MyFinanceApp - Kotlin - Accounts Fragment

+++ 2022-2023/sourcecode/lab/lab5/MyFinanceAppKotlin/app/src/main/java/pl/edu/uwr/pum/myfinanceappkotlin/data/Accounts.kt
@@ -1,10 +1,8 @@
 package pl.edu.uwr.pum.myfinanceappkotlin.data
 
-import android.graphics.Color
-
 data class Accounts(
     val name: String,
     val number: String,
-    val amount: Int,
+    val amount: Double,
     val color: Int
 )
